:recycle: reshuffle code to make it available elsewhere
