+++ client/ayon_core/pipeline/traits/content.py
@@ -9,6 +9,8 @@ from dataclasses import dataclass
 from pathlib import Path  # noqa: TCH003
 from typing import ClassVar, Generator, Optional
 
+from clique import assemble
+
 from .representation import Representation
 from .temporal import FrameRanged, Handles, Sequence
 from .trait import (
@@ -17,7 +19,6 @@ from .trait import (
     TraitValidationError,
 )
 from .two_dimensional import UDIM
-from .utils import get_sequence_from_files
 
 
 @dataclass
@@ -212,7 +213,7 @@ class FileLocations(TraitBase):
                 representation.
 
         """
-        tmp_frame_ranged: FrameRanged = get_sequence_from_files(
+        tmp_frame_ranged: FrameRanged = self.get_sequence_from_files(
                     [f.file_path for f in self.file_paths])
 
         frames_from_spec: list[int] = []
@@ -347,6 +348,46 @@ class FileLocations(TraitBase):
         frame_end_with_handles = frame_end + frame_end_handle
 
         return frame_start_with_handles, frame_end_with_handles
+
+    @staticmethod
+    def get_sequence_from_files(paths: list[Path]) -> FrameRanged:
+        """Get the original frame range from files.
+
+        Note that this cannot guess frame rate, so it's set to 25.
+        This will also fail on paths that cannot be assembled into
+        one collection without any reminders.
+
+        Args:
+            paths (list[Path]): List of file paths.
+
+        Returns:
+            FrameRanged: FrameRanged trait.
+
+        Raises:
+            ValueError: If paths cannot be assembled into one collection
+
+        """
+        cols, rems = assemble([path.as_posix() for path in paths])
+        if rems:
+            msg = "Cannot assemble paths into one collection"
+            raise ValueError(msg)
+        if len(cols) != 1:
+            msg = "More than one collection found"
+            raise ValueError(msg)
+        col = cols[0]
+
+        sorted_frames = sorted(col.indexes)
+        # First frame used for end value
+        first_frame = sorted_frames[0]
+        # Get last frame for padding
+        last_frame = sorted_frames[-1]
+        # Use padding from a collection of the last frame lengths as string
+        # padding = max(col.padding, len(str(last_frame)))
+
+        return FrameRanged(
+            frame_start=first_frame, frame_end=last_frame,
+            frames_per_second="25.0"
+        )
 
 
 @dataclass
