+++ client/ayon_core/plugins/publish/integrate_hero_version_with_traits.py
@@ -1,11 +1,18 @@
 """Integrate Hero version with representation traits."""
 from __future__ import annotations
+
+import json
+import os
+from pprint import pformat
+import shutil
 import copy
-from ayon_core.pipeline.publish import get_trait_representations
 from ayon_core.pipeline.traits import (
-    Persistent,
     Representation,
-    TraitBase,
+    get_transfers_from_representations,
+)
+from ayon_core.lib.file_transaction import (
+    DuplicateDestinationError,
+    FileTransaction,
 )
 from ayon_api.operations import (
     OperationsSession,
@@ -14,7 +21,7 @@ from ayon_api.operations import (
 
 import pyblish.api
 import ayon_api
-from ayon_api.utils import create_entity_id
+
 from typing import TYPE_CHECKING, Optional
 
 from ayon_core.pipeline.publish import (
@@ -81,6 +88,8 @@ class IntegrateHeroVersionTraits(
     ignored_representation_names: set[str] = set()
     log: "logging.Logger"
 
+    use_hardlinks = False
+
     def process(self, instance: pyblish.api.Instance) -> None:
         """Integrate Hero version with representation traits.
 
@@ -105,9 +114,7 @@ class IntegrateHeroVersionTraits(
                 f"'{template_key}' template key!")
             return
 
-        self.log.debug(
-            f"'hero' template check was successful. '{hero_template}'"
-        )
+        hero_publish_dir = self.get_publish_dir(instance, template_key)
 
         src_version_entity = instance.data.get("versionEntity")
 
@@ -123,10 +130,15 @@ class IntegrateHeroVersionTraits(
             self.log.warning("Version 0 cannot have hero version. Skipping.")
             return
 
-        # Current version
+        # Current hero version data
         old_version, old_repres = self.current_hero_entities(
             project_name, src_version_entity
         )
+
+        # old representations are coming from already existing hero version
+        # new representations are coming from current version that is
+        # being published
+
 
         op_session = OperationsSession()
 
@@ -139,6 +151,7 @@ class IntegrateHeroVersionTraits(
             attribs=copy.deepcopy(src_version_entity["attrib"]),
             entity_id=entity_id,
         )
+        self.log.debug(f"Prepared hero version entity: {new_hero_version}")
 
         if old_version:
             self.log.debug("Replacing old hero version.")
@@ -161,9 +174,13 @@ class IntegrateHeroVersionTraits(
         instance.data["heroVersionEntity"] = new_hero_version
 
         # get published representations with traits for the version
-        repre_entities = ayon_api.get_representations(
+        repre_entities = list(ayon_api.get_representations(
             project_name=project_name,
-            version_ids={new_hero_version["id"]})
+            version_ids={src_version_entity["id"]}))
+
+        self.log.debug(
+            f"Found {len(repre_entities)} representations for hero version."
+        )
 
         if not repre_entities:
             msg = (
@@ -174,9 +191,7 @@ class IntegrateHeroVersionTraits(
             self.log.error(msg)
             raise KnownPublishError(msg)
 
-
         # Separate old representations into `to replace` and `to delete`
-
         inactive_old_repres_by_name = {}
         old_repres_by_name = {}
         for repre in old_repres:
@@ -187,21 +202,138 @@ class IntegrateHeroVersionTraits(
                 inactive_old_repres_by_name[low_name] = repre
 
         old_repres_to_replace = {}
-        old_repres_to_delete = {}
 
-        for repre in filtered_representations:
-            repre_name_low = repre.name.lower()
+
+        for repre in repre_entities:
+            repre_name_low = repre["name"].lower()
             if repre_name_low in old_repres_by_name:
                 old_repres_to_replace[repre_name_low] = (
                     old_repres_by_name.pop(repre_name_low)
                 )
 
-        if old_repres_by_name:
-            old_repres_to_delete = old_repres_by_name
+        old_repres_to_delete = old_repres_by_name or {}
+        backup_hero_publish_dir = None
+        if os.path.exists(hero_publish_dir):
+            backup_hero_publish_dir = self._backup_hero_version_dir(
+                hero_publish_dir)
 
-        if old_repres_by_name:
-            old_repres_to_delete = old_repres_by_name
+        for repe in repre_entities:
+            self.log.debug(f"representation: {pformat(repe, indent=4)}")
 
+        representations = [
+            Representation.from_dict(
+                name=repre["name"],
+                representation_id=repre["id"],
+                trait_data=json.loads(repre["traits"])
+            )
+            for repre in repre_entities
+            if repre["name"] not in self.ignored_representation_names and repre["traits"]
+        ]
+
+        self.log.debug(
+            "Prepared representations for hero version: %s",
+            [repre.name for repre in representations]
+        )
+        self.log.debug(f"Hero template: {hero_template}")
+
+        transfers = get_transfers_from_representations(
+            instance, hero_template, representations)
+
+        file_transactions = FileTransaction(
+            log=self.log,
+            # Enforce unique transfers
+            allow_queue_replacements=False
+        )
+        mode = FileTransaction.MODE_COPY
+        if self.use_hardlinks:
+            mode = FileTransaction.MODE_HARDLINK
+
+        try:
+            for transfer in transfers:
+                self.log.debug(
+                    "Transferring file: %s -> %s",
+                    transfer.source,
+                    transfer.destination
+                )
+                file_transactions.add(
+                    transfer.source.as_posix(),
+                    transfer.destination.as_posix(),
+                    mode=mode,
+                )
+            file_transactions.process()
+        except DuplicateDestinationError as e:
+            msg = (
+                "Multiple representations are trying to transfer files to "
+                "the same destination. This is not allowed because it can "
+                "cause conflicts and unintended overwrites. Please check the "
+                "representations and their traits to ensure they are unique."
+            )
+            self.log.error(msg)
+            raise KnownPublishError(msg) from e
+        except Exception as e:
+            msg = (
+                "An error occurred during file transfer for hero version. "
+                "Please check the logs for more details."
+            )
+            self.log.error(msg)
+            raise KnownPublishError(msg) from e
+        finally:
+            file_transactions.finalize()
+
+    def _backup_hero_version_dir(self, hero_publish_dir: str) -> str:
+        """Backup current hero version publish directory.
+
+        Args:
+            hero_publish_dir (str): The path to current hero
+                version publish directory.
+        """
+        backup_hero_publish_dir = f"{hero_publish_dir}.BACKUP"
+        # max backup dirs present
+        max_idx = 10
+        idx = 0
+        _backup_hero_publish_dir = backup_hero_publish_dir
+        while os.path.exists(_backup_hero_publish_dir):
+            self.log.debug(
+                "Backup folder already exists. "
+                f'Trying to remove "{_backup_hero_publish_dir}"'
+            )
+
+            try:
+                shutil.rmtree(_backup_hero_publish_dir)
+                backup_hero_publish_dir = _backup_hero_publish_dir
+                break
+            except Exception:
+                self.log.info(
+                    "Could not remove previous backup folder. "
+                    "Trying to add index to folder name."
+                )
+
+            _backup_hero_publish_dir = (
+                    backup_hero_publish_dir + str(idx)
+            )
+            if not os.path.exists(_backup_hero_publish_dir):
+                backup_hero_publish_dir = _backup_hero_publish_dir
+                break
+
+            if idx > max_idx:
+                msg = (
+                    "Backup folders are fully occupied "
+                    f'to max index "{max_idx}"'
+                )
+                raise AssertionError(msg)
+            idx += 1
+
+        self.log.debug(f'Backup folder path is \"{backup_hero_publish_dir}\"')
+        try:
+            os.rename(hero_publish_dir, backup_hero_publish_dir)
+        except PermissionError as e:
+            msg = (
+                "Could not create hero version because it is not "
+                "possible to replace current hero files."
+            )
+            raise AssertionError(msg) from e
+
+        return backup_hero_publish_dir
 
     @staticmethod
     def version_from_representations(
@@ -244,9 +376,10 @@ class IntegrateHeroVersionTraits(
         anatomy_data = instance.data["anatomyData"]
         task_info = anatomy_data.get("task") or {}
         host_name = instance.context.data["hostName"]
-        product_base_type = instance.data.get("productBaseType")
-        if not product_base_type:
-            product_base_type = instance.data["productType"]
+        product_base_type = (
+                instance.data.get("productBaseType")
+                or instance.data["productType"]
+        )
 
         return get_publish_template_name(
             project_name,
@@ -286,3 +419,36 @@ class IntegrateHeroVersionTraits(
             project_name, version_ids={hero_version["id"]}
         ))
         return hero_version, hero_repres
+
+    def get_publish_dir(
+            self,
+            instance: pyblish.api.Instance,
+            template_key: str) -> str:
+        """Get publish directory for hero version.
+
+        Args:
+            instance (pyblish.api.Instance): The instance to get data from.
+            template_key (str): The template key to use for hero template.
+
+        Returns:
+            str: The path to publish directory for hero version.
+
+        """
+        anatomy = instance.context.data["anatomy"]
+        template_data = copy.deepcopy(instance.data["anatomyData"])
+
+        if "originalBasename" in instance.data:
+            template_data.update({
+                "originalBasename": instance.data.get("originalBasename")
+            })
+
+        template_obj = anatomy.get_template_item(
+            "hero", template_key, "directory"
+        )
+        publish_folder = os.path.normpath(
+            template_obj.format_strict(template_data)
+        )
+
+        self.log.debug(f'hero publish dir: "{publish_folder}"')
+
+        return publish_folder
